requirements.txt
gradio==3.40.1
langchain
langchain-community
faiss-cpu
sentence-transformers
PyMuPDF
huggingface_hub